fix reference to `JIT` by moving `CompileHostModule` to Runtime component
