+++ src/enzyme_ad/jax/Runtime/BUILD
@@ -18,5 +18,7 @@ cc_library(
         "@llvm-project//llvm:OrcJIT",
         "@llvm-project//llvm:OrcTargetProcess",
         "@llvm-project//llvm:Support",
+        "@llvm-project//mlir:IR",
+        "@llvm-project//mlir:ToLLVMIRTranslation",
     ],
 )
